+++ tests/stats/test_prediction_bias.py
@@ -6,7 +6,7 @@ import pytest
 
 from secretflow import reveal
 from secretflow.data import FedNdarray, PartitionWay
-from secretflow.data.base import Partition
+from secretflow.data.base import partition
 from secretflow.data.vertical import VDataFrame
 from secretflow.stats import prediction_bias_eval
 from secretflow.stats.core.prediction_bias_core import PredictionBiasBucketMethod
@@ -25,7 +25,7 @@ def prod_env_and_data(sf_production_setup_devices):
     )
     y_actual = VDataFrame(
         partitions={
-            sf_production_setup_devices.alice: Partition(
+            sf_production_setup_devices.alice: partition(
                 data=sf_production_setup_devices.alice(lambda x: x)(
                     y_actual_pd_dataframe
                 )