+++ CHANGELOG.md
@@ -12,6 +12,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 `Fixed` for any bug fixes.
 `Security` in case of vulnerabilities.
 
+## [1.2.0.dev230926] - 2023-9-26
+
+### Changed
+- component: expose cross_silo_comm_backend option for secretflow init.
+
+### Fixed
+- SPU device: rm invalid use_link option.
+
+## [1.2.0.dev230925] - 2023-9-25
+
+### Changed
+- update psi benchmark.
+
+### Fixed
+- fix tls on kuscia.
+
+## [1.2.0.dev230918] - 2023-9-18
+
+### Added
+- support GLM model transition from MPC version to Federated version.
+- add sparse, quantized and mixed compressor.
+- add polars backend for dataframe and SL model.
+- add DisPFL adn VFGNN example.
+- add cross_silo_comm_backend option to SFClusterConfig.RayFedConfig.
+
+### Fixed
+- fix dataset build in SL.
+
 ## [1.1.0b0] - 2023-9-6
 
 ### Added
@@ -30,7 +58,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Stateful task for teeu
-- GPU support for torch fl model
 - docs: DeepFM translation
 
 ### Changed